Add: Tab information field now shows information about the selected desc.
If any owned vehicles is selected as well, it will show additional
information, such as reassigned classes, resale value etc.
Fix: The information now lives in the first tab, currently called
"general information"
Fix: If a vehicle was outside the map, for some reason, it crashed when
attempting to find a city. Instead, it will now show the coordinates of
the vehicle. Also when sorting by location, it now checks if the vehicle
is actually on the map.